+++ ijkmedia/ijkplayer/ijkavutil/ijkstl.h
@@ -20,15 +20,17 @@
 #ifndef IJKAVUTIL_IJKSTL_H
 #define IJKAVUTIL_IJKSTL_H
 
+#include <stdint.h>
+
 void* ijk_map_create();
-void ijk_map_put(void *data, int key, void *value);
-void* ijk_map_get(void *data, int key);
-int ijk_map_remove(void *data, int key);
+void ijk_map_put(void *data, int64_t key, void *value);
+void* ijk_map_get(void *data, int64_t key);
+int ijk_map_remove(void *data, int64_t key);
 int ijk_map_size(void *data);
 int ijk_map_max_size(void *data);
 void* ijk_map_index_get(void *data, int index);
 void ijk_map_traversal_handle(void *data, int (*enu)(void *elem));
-int ijk_map_get_min_key(void *data);
+int64_t ijk_map_get_min_key(void *data);
 void ijk_map_clear(void *data);
 void ijk_map_destroy(void *data);
